Fixes #54
Move the contents of `fallibleSyntax.scala` to `syntax.scala`
