+++ ducktape/src/main/scala/io/github/arainko/ducktape/syntax.scala
@@ -19,3 +19,22 @@ extension [Source](value: Source) {
     Source: Mirror.ProductOf[Source]
   ): Func.Return = Transformations.via(value, function)
 }
+
+extension [F[+x], Source](value: Source)(using F: Transformer.Mode[F]) {
+
+  inline def fallibleTo[Dest](using transformer: Transformer.Fallible[F, Source, Dest]): F[Dest] =
+    transformer.transform(value)
+
+  inline def fallibleVia[Func](inline function: Func)(using
+    Func: FunctionMirror[Func]
+  )(using Source: Mirror.ProductOf[Source]): F[Func.Return] =
+    inline F match {
+      case given fallible.Mode.FailFast[F] =>
+        Transformations.failFastVia[F, Source, Func](function)(value)
+      case given fallible.Mode.Accumulating[F] =>
+        Transformations.accumulatingVia[F, Source, Func](function)(value)
+      case other =>
+        Errors.cannotDetermineTransformationMode
+    }
+
+}
